Develop

See merge request ai_track/class_06/web_project/team06/6team_web!16

+++ server/services/products.js
@@ -1,17 +1,17 @@
-const { Product } = require('../db/model');
+const { Product, Category } = require('../db/model');
 
 export default class ProductsService {
   static async getAllProduct(title) {
-    const products = await Product.find();
-    if (!title) {
-      return products;
-    } else {
+    if (title) {
       const category = await Category.find({ title });
       const products = await Product.find({ categoryId: category.id });
       if (!products) {
         throw new Error('존재하지 않는 제품입니다');
       }
       return products;
+    } else {
+      const allProducts = await Product.find();
+      return allProducts;
     }
   }
 
@@ -31,6 +31,7 @@ export default class ProductsService {
     imageKey,
     inventory,
     price,
+    role,
   }) {
     if (
       !title ||
@@ -43,33 +44,67 @@ export default class ProductsService {
     ) {
       throw new Error('필수 항목이 모두 채워지지 않았습니다.');
     }
-    const createdProduct = await Product.create({
-      title,
-      categoryId,
-      shortDescription,
-      detailDescription,
-      imageKey,
-      inventory,
-      price,
-    });
-    return createdProduct;
+    if (role === 'ADMIN') {
+      const createdProduct = await Product.create({
+        title,
+        categoryId,
+        shortDescription,
+        detailDescription,
+        imageKey,
+        inventory,
+        price,
+      });
+      return createdProduct;
+    } else {
+      throw new Error('권한이 없습니다.');
+    }
   }
 
-  static async updateProductById(id) {
-    const updatedProduct = await Product.findByIdAndUpdate(id, req.body, {
-      new: true,
-    });
-    if (!updatedProducts) {
-      throw new Error('존재하지 않는 제품 입니다');
+  static async updateProductById({
+    id,
+    title,
+    categoryId,
+    shortDescription,
+    detailDescription,
+    imageKey,
+    inventory,
+    price,
+    role,
+  }) {
+    if (role === 'ADMIN') {
+      const updatedProduct = await Product.findByIdAndUpdate(
+        id,
+        {
+          ...(title && { title }),
+          ...(categoryId && { categoryId }),
+          ...(shortDescription && { shortDescription }),
+          ...(detailDescription && { detailDescription }),
+          ...(imageKey && { imageKey }),
+          ...(inventory && { inventory }),
+          ...(price && { price }),
+        },
+        {
+          new: true,
+        },
+      );
+      if (!updatedProduct) {
+        throw new Error('존재하지 않는 제품 입니다');
+      }
+      return updatedProduct;
+    } else {
+      throw new Error('권한이 없습니다.');
     }
-    return updatedProduct;
   }
 
-  static async deleteProductById(id) {
-    const deletedProduct = await Product.findByIdAndDelete(id);
-    if (!deletedProduct) {
-      throw new Error('존재하지 않는 제품 입니다');
+  static async deleteProductById(id, role) {
+    if (role === 'ADMIN') {
+      const deletedProduct = await Product.findByIdAndDelete(id);
+      if (!deletedProduct) {
+        throw new Error('존재하지 않는 제품 입니다');
+      }
+      return deletedProduct;
+    } else {
+      throw new Error('권한이 없습니다.');
     }
-    return deletedProduct;
   }
 }
